refactor(code): trim resumed diff restoration
